selftests/bpf: Add test for freplace program with write access
Summary:
selftests/bpf: Add test for freplace program with write access

This adds a selftest that tests the behavior when a freplace target program
attempts to make a write access on a packet. The expectation is that the read or write
access is granted based on the program type of the linked program and
not itself (which is of type, for e.g., BPF_PROG_TYPE_EXT).

This test fails without the associated patch on the verifier.


(cherry picked from commit 6dc03dc71387e1dc65cf14efb49e5cf7062a2d46)

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Udip Pant <udippant@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/20200825232003.2877030-3-udippant@fb.com