selftests/bpf: Test for checking return code for the extended prog
Summary:
selftests/bpf: Test for checking return code for the extended prog

This adds test to enforce same check for the return code for the extended prog
as it is enforced for the target program. It asserts failure for a
return code, which is permitted without the patch in this series, while
it is restricted after the application of this patch.


(cherry picked from commit 50d19736aff497a4c25ec7e36375195bfd8570cd)

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Udip Pant <udippant@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/20200825232003.2877030-4-udippant@fb.com