FB only: efi: add tpm12 cmdline parameter
Adding 'tpm12' to the cmdline forces the system to only
attempt the UEFI TPM 1.2 protocol during booting.  Normal
procedure involves trying TPM 2.0 first, and if that fails,
then fall back to TPM 1.2.

The original kernel behavior is restored when no cmdline
override is present.

This replaces the previous two TPM changes, which changed
the default to 1.2, and addedd a 'notpm' flag which
bypassed retrieving the TPM event logs completely.