mm: don't force readahead with a NULL filp
Summary:
mm: don't force readahead with a NULL filp

Jens added something that forces the page cache readahead if we're
congested, but this particular helper uses the filp for the readahead
state.  The problem is we can pass in NULL for filp, which btrfs does,
and if we're under congestion we'll force the readhead via
force_page_cache_readahead and NULL pointer deref on the filp.  Fix this
by not doing the force ra if we don't have a filp.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <jbacik@fb.com>