readahead: don't do readahead if it's disabled and we don't have a filp
Summary:
readahead: don't do readahead if it's disabled and we don't have a filp

In fixing the null pointer deref for the !filp case I made it so we
always do readahead in the !filp case, which isn't awesome either.  Fix
the logic to bail if ra is disabled and we have !filp.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <jbacik@fb.com>